Unify-transformer_lm_megatron-and-transformer_lm (#634)
* move to transformer_lm_mp.py

* remove transformer_lm_mp

* add arch transformer_lm_gpt and transformer_lm_gpt2_tiny

* update arch to model_parallel_transformer_lm

* rename to transformer_lm.py

* remove transformer_decoder_mp.py

* remove transformer_decoder_layer_mp.py

* fix lint errors

---------
